+++ python/scripts/mathematische-operatoren.py
@@ -1,5 +1,5 @@
-zahl = int(input("Bitte gib eine Zahl ein: "))
-vergleich = int(input("Bitte gib eine Zahl ein, mit der du deine Zahl vergleichen möchtest: "))
+zahl = float(input("Bitte gib eine Zahl ein: "))
+vergleich = float(input("Bitte gib eine Zahl ein, mit der du deine Zahl vergleichen möchtest: "))
 
 if zahl < vergleich:
     print("Deine Zahl ist kleiner.")
@@ -8,8 +8,8 @@ elif zahl > vergleich:
 else:
     print("Du hast die selbe Zahl eingegeben.")
     
-zahl2 = int(input("Bitte gib eine zweite Zahl ein: "))
-vergleich2 = int(input("Bitte gib eine zweite Zahl ein, mit der du deine (zweite) Zahl vergleichen möchtest: "))
+zahl2 = float(input("Bitte gib eine zweite Zahl ein: "))
+vergleich2 = float(input("Bitte gib eine zweite Zahl ein, mit der du deine (zweite) Zahl vergleichen möchtest: "))
 
 if zahl <= vergleich:
     print("Deine Zahl ist kleiner oder gleich.")
